+++ pipeline/execution.py
@@ -1,3 +1,4 @@
+import os
 import subprocess
 import sys
 import time
@@ -68,7 +69,7 @@ def execute_dag_serially(dag, env, config):
 
             path = _preprocess_task(id_, dag, env, config)
 
-            _ = _execute_task(id_, path, config["_is_debug"])
+            _ = _execute_task(id_, path, config)
 
             _process_task_target(id_, dag, config)
 
@@ -104,7 +105,7 @@ def execute_dag_parallelly(dag, env, config):
 
                 path = _preprocess_task(id_, dag, env, config)
 
-                future = executor.submit(_execute_task, id_, path, config["_is_debug"])
+                future = executor.submit(_execute_task, id_, path, config)
                 future.add_done_callback(lambda x: t.update())
                 submitted_tasks[id_] = future
 
@@ -117,7 +118,7 @@ def execute_dag_parallelly(dag, env, config):
 
             # Check for exceptions.
             exceptions = [
-                future.exception().message
+                str(future.exception())
                 for future in submitted_tasks.values()
                 if future.exception()
             ]
@@ -137,10 +138,14 @@ def execute_dag_parallelly(dag, env, config):
 def _collect_unfinished_tasks(dag, env, config):
     """Collect unfinished tasks.
 
-    Iterate over topological sorted nodes in the DAG. If the node is a task, compare the
-    hashes of all dependencies and targets. If the hashes do not match, add the task to
-    the set of unfinished tasks. After that, go through the whole list of descendants of
-    the task and mark all tasks among them as unfinished, too.
+    Iterate over topological sorted nodes in the DAG. If the node is a task, do the
+    following.
+
+    1. If the task is marked to be always executed, add it to the set.
+    2. Otherwise, compare the hashes of all dependencies and targets. If the hashes do
+       not match, add the task to the set of unfinished tasks. After that, go through
+       the whole list of descendants of the task and mark all tasks among them as
+       unfinished, too.
 
     Parameters
     ----------
@@ -154,18 +159,34 @@ def _collect_unfinished_tasks(dag, env, config):
     """
     unfinished_tasks = set()
     for id_ in nx.topological_sort(dag):
-        if dag.nodes[id_]["_is_task"] and id_ not in unfinished_tasks:
-            have_same_hashes = compare_hashes_of_task(id_, env, dag, config)
-            if not have_same_hashes:
+        if dag.nodes[id_]["_is_task"]:
+            if dag.nodes[id_].get("run_always", False):
                 unfinished_tasks.add(id_)
-                for descendant in nx.descendants(dag, id_):
-                    if dag.nodes[descendant]["_is_task"]:
-                        unfinished_tasks.add(descendant)
+            else:
+                have_same_hashes = compare_hashes_of_task(id_, env, dag, config)
+                if not have_same_hashes:
+                    unfinished_tasks.add(id_)
+                    for descendant in nx.descendants(dag, id_):
+                        if dag.nodes[descendant]["_is_task"]:
+                            unfinished_tasks.add(descendant)
 
     return unfinished_tasks
 
 
 def _compute_padding_to_prevent_task_description_from_moving(unfinished_tasks):
+    """Compute the padding to have task descriptions with the same length.
+
+    Some task names are longer than others. The tqdm progress bar would be constantly
+    adjusting if more space is available. Instead, we compute the length of the longest
+    task name and add whitespace to the right.
+
+    Example
+    -------
+    >>> unfinished_tasks = ["short_name", "long_task_name"]
+    >>> _compute_padding_to_prevent_task_description_from_moving(unfinished_tasks)
+    14
+
+    """
     len_task_names = list(map(len, unfinished_tasks))
     padding = max(len_task_names) if len_task_names else 0
 
@@ -175,8 +196,8 @@ def _compute_padding_to_prevent_task_description_from_moving(unfinished_tasks):
 def _preprocess_task(id_, dag, env, config):
     file = render_task_template(id_, dag.nodes[id_], env, config)
 
-    if "produces" in dag.nodes[id_]:
-        Path(dag.nodes[id_]["produces"]).parent.mkdir(parents=True, exist_ok=True)
+    for target in ensure_list(dag.nodes[id_].get("produces", [])):
+        Path(target).parent.mkdir(parents=True, exist_ok=True)
 
     if dag.nodes[id_]["template"].endswith(".py"):
         path = Path(config["hidden_task_directory"], id_ + ".py")
@@ -194,18 +215,23 @@ def _preprocess_task(id_, dag, env, config):
     return path
 
 
-def _execute_task(id_, path, is_debug):
+def _execute_task(id_, path, config):
     if path.suffix == ".py":
+        environment = _patch_subprocess_environment(config)
+
         try:
-            subprocess.run(["python", str(path)], check=True)
+            subprocess.run(["python", str(path)], check=True, env=environment)
 
         except subprocess.CalledProcessError as e:
             message = _format_exception_message(id_, path, e)
-            if is_debug:
+            if config["_is_debug"]:
                 click.echo(message)
                 click.echo("Rerun the task to enter the debugger.")
+
                 subprocess.run(
-                    ["python", "-m", "pdb", "-c", "continue", str(path)], check=True,
+                    ["python", "-m", "pdb", "-c", "continue", str(path)],
+                    check=True,
+                    env=environment,
                 )
                 sys.exit("### Abort build.")
             else:
@@ -240,6 +266,22 @@ def _process_task_target(id_, dag, config):
     ]
     if missing_targets:
         raise FileNotFoundError(
-            f"Target(s) '{missing_targets}' was(were) not produced by task '{id_}'."
+            f"Target(s) {missing_targets} was(were) not produced by task '{id_}'."
         )
     save_hash_of_task_target(id_, dag, config)
+
+
+def _patch_subprocess_environment(config):
+    """Patch the environment of the subprocess.
+
+    The problem is that task files are rendered and, then, stored in the hidden build
+    directory. This would prohibit imports if we did not add the project root to the
+    `PYTHONPATH`.
+
+    """
+    env = os.environ.copy()
+    env["PYTHONPATH"] = (
+        str(Path(config["project_directory"])) + ";" + env.get("PYTHONPATH", "")
+    )
+
+    return env