+++ docs/release_notes.rst
@@ -1,10 +1,28 @@
-Changes
-=======
+Release notes
+=============
 
 This is a record of all past ``pipeline`` releases and what went into them in reverse
 chronological order. We follow `semantic versioning <https://semver.org/>`_ and all
 releases are available on `Anaconda.org
 <https://anaconda.org/opensourceeconomics/pipeline>`_.
+
+
+0.0.6 - 2020-
+-------------
+
+- Add ``run_always`` to the task dictionary to always execute a task (:gh:`12`).
+- Add project root directory to the ``PYTHONPATH`` so that every task can import with
+  ``from src.... import ...`` (:gh:`13`).
+- Allow to have directories as task dependencies and targets (:gh:`15`).
+
+
+0.0.5 - 2020-04-26
+------------------
+
+- Added ``pydot`` as a package dependency  and added all information in ``config`` to
+  render a task template (:gh:`11`).
+- Remove references to respy (:gh:`8`).
+- Fix codecov (:gh:`7`).
 
 
 0.0.4 - 2020-04-06